refactor: useAddOrUpdateEvent 추출
